### 2023-01-30 #5111 @faroit
@tuomaseerola 👋  hanks for your submission to JOSS! I will be serving as the editor for this. We'll use this issue to identify reviewers and resolve technical issues.

Tt would be great to have reviewers with domain knowledge who also have experience with frameworks used in the MIR community... I have a few potential reviewers in mind but @tuomaseerola, I am also interested in your input on potential reviewers, so any suggestions you have are also welcome (especially with R knowledge)

For this, please check the JOSS [COI policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy) and only mention them with a space between the @ and their handle (so you would refer to me as "@ faroit").
