### 2023-02-07 #5111 @faroit
👋  @magdalenafuentes @sstober @jacobyn @pmcharrison @ttroes @lminer - would any of you be willing to review this submission for JOSS? The JOSS review process takes place on GitHub and focuses on the software and a short paper. We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: <https://joss.readthedocs.io/en/latest/reviewer_guidelines.html>

The software under review is [onsetsync](https://github.com/tuomaseerola/onsetsync).
