### 2023-02-14 #5111 @magdalenafuentes
Hey @faroit thanks for reaching out. The work does look super interesting, but I don't have much experience in R so I feel unqualified to review it. May I suggest @mrocamora as a reviewer instead? He has more experience in R and he's the main author of [Carat](https://github.com/mrocamora/carat), a Python library for computer-aided rhythm analysis with similar features to the ones proposed here.  I saw that Martín is in the acknowledgements so not sure if that would be a conflict of interest, @tuomaseerola might be able to answer that.
