### 2023-02-22 #5111 @hendriks73
Hey @faroit! I would love to review this. Looks very interesting! Unfortunately, I don't know R and am way too busy changing diapers. Yes—that's what I do full time now! 😉

Perhaps you want to contact @julian-urbano (who is is an R expert).
Good luck and sorry I'm not the right person for this.
