### 2023-04-21 #5395 @editorialbot
[REVIEW]: onsetsync: An R Package for Onset Synchrony Analysis
**Submitting author:** @tuomaseerola (<a href="http://orcid.org/0000-0002-2896-929X">Tuomas Eerola</a>)
**Repository:** https://github.com/tuomaseerola/onsetsync
**Branch with paper.md** (empty if default branch): master
**Version:** 0.5.1
