### 2023-04-23 #5395 @faroit
👋🏼 @tuomaseerola @pmcharrison, @mrocamora this is the review thread for the paper. All of our communications will happen here from now on.

```@editorialbot generate my checklist```
