### 2023-04-23 tuomaseerola/onsetsync#1 @faroit
[JOSS] fix DOIs
Hi,

it appears that the DOIs for a couple of references are incorrect. Can you please adjust them so that they render correctly?

this is part of the https://github.com/openjournals/joss-reviews/issues/5395 review
