### 2023-04-25 #5395 @pmcharrison
Typos in this sentence:

> These collection contain tens of hours annotated and verified
