### 2023-04-25 #5395 @pmcharrison
Typo here:

> a processes occurring within 100-2000 ms timescales
