### 2023-04-25 #5395 @pmcharrison
Wrong citation format here:

> A number of different measures of synchrony have been proposed. We follow the definitions
54 from (Clayton et al., 2020, p. 161)
