### 2023-04-25 #5395 @pmcharrison
Eq. 1: sum symbol should say `i`
