### 2023-04-25 #5395 @pmcharrison
Eq. 2: why not use `d_i` abbreviation here?
