### 2023-04-25 #5395 @pmcharrison
Eq. 3: `T` has not been defined yet
