### 2023-04-25 #5395 @pmcharrison
Eq. 4: why has `T` now been changed to `n`? Same in Eq. 5
