### 2023-04-25 #5395 @pmcharrison
Eq. 4: `i=i` should be `i=1`
