### 2023-04-25 #5395 @pmcharrison
Eq. 5: isn't this the same as d bar, used in Eq. 2? If so, use the same symbol for both?
