### 2023-04-25 #5395 @pmcharrison
Eq. 8: `i=i` should be `i=1`.

I'm also not sure what the subscript `r` means here.
