### 2023-04-25 #5395 @pmcharrison
> which is usually the relevant approach in music sense.

doesn't seem like idiomatic English.
