### 2023-04-25 #5395 @pmcharrison
Wrong citation formats:

> see (Schlüter & Böck, 2014) and (Böck et al., 2016), and for a full workflow of how to combine onset detection and annotation of the musical information, see (Clayton, Tarsitani, et al., 2021).
