### 2023-04-25 #5395 @pmcharrison
Table 2: it seems unfortunate that SD is used as a column name, because most readers will think this denotes standard deviation. Probably it's too late to change now though
