### 2023-04-25 #5395 @pmcharrison
Typo:

> As a broad overview, we can use the function plot_by_beat visualise the asynchrony relative to
