### 2023-04-25 #5395 @pmcharrison
Regarding this checklist point,

> Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

I could not see any automated tests, but maybe I am missing them?
