### 2023-04-25 #5395 @pmcharrison
Regarding this checklist point,

>  Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

I could not see any, but maybe I am missing them?
