### 2023-04-25 #5395 @pmcharrison
@tuomaseerola I have gone through the paper and it generally looks good, I think the functionality is appealing and I particularly like the visualisation side. I can see that the pipeline would work very well for future researchers in that the input data format is quite general and once you have in that format it's very easy to apply your code to it.

The paper is a bit long for JOSS format but I think it works to have those detailed examples (I didn't check whether it exceeds the 1,000 word limit, it's probably close). There were quite a few typographical issues and I think the draft could do with some detailed proofing to complement what I was able to spot. I have made detailed individual comments above.
