### 2023-04-27 #5395 @mrocamora
Regarding COIs, I declare that I participated in the Interpersonal Entrainment in Music Performance (IEMP) project, which took place from 2016 to 2018. I worked on preparing Uruguayan candombe data for the IEMP collection in collaboration with Luis Jure, Simone Tarsitani, and Martin Clayton. I also discussed and explored approaches for precise onset extraction from audio recordings with Martin Clayton and Tuomas Eerola. However, I think I can make an impartial assessment of the work presented to JOSS, whose focus is the analysis of onsets to assess musical synchrony.
