### 2023-05-09 #5395 @mrocamora
> Here we first define the onset time differences, 𝑑, ...

Then, in the definition of the precision measures, the term "asynchronies" is used for d, instead of "onset time differences". So I suggest: "Here we first define the onset time differences (or asynchronies), 𝑑, ..."

"Mean of all unsigned onset time difference values ..." seems more appropriate.
