### 2023-05-09 #5395 @mrocamora
Eq. 4: Summation starts at $i=1$
