### 2023-05-09 #5395 @mrocamora
Eq. 5: Seems that $\bar{d}$ should be used instead of $\hat{s_p}$ (as in Eq. 2). I don't see where the $p$ comes from.
