### 2023-05-09 #5395 @mrocamora
> Mean relative asynchrony : mean position of an instrument’s onsets relative to average position $(\hat{\omega})$ of the group, ...

The position of an instrument's onsets was previously notated as $I$ ($I_{1,i} and $I_{2,i}$). I suggest using a consistent notation for onsets. Besides, I also suggest using $\bar{\omega}$ instead of $\hat{\omega}$, as previously done in Eq. 2.
