### 2023-05-09 #5395 @mrocamora
> and the relative onset time $w$ is

Should be $v$ instead of $w$.
