### 2023-05-10 #5395 @mrocamora
Eq. 8: I suggest using $\bar{s_r}$ instead of $\hat{s_r}$.
