### 2023-05-10 #5395 @mrocamora
> which is usually the relevant approach in music sense.

This assertion could be backed up by citing some relevant references.
