### 2023-05-10 #5395 @mrocamora
>  (e.g. Librosa, Essentia, or MIR toolbox for Matlab, or Sonic Visualiser using well-known onset detection algorithms.

There is an extra "or" and the closing parenthesis is missing.
