### 2023-05-10 #5395 @mrocamora
>  (with precision of 6 digits, ...

Closing parenthesis is missing.
