### 2023-05-10 #5395 @mrocamora
Sec 1.6 Onset summaries

Note that you need to load ``library(dplyr)'' to run the example code, as it internally uses the ``ungroup'' function.
