### 2023-05-10 #5395 @mrocamora
> As a broad overview, we can use the function plot_by_beat visualise the asynchrony

to visualise
