### 2023-05-10 #5395 @mrocamora
> Overall this suggests that these two instruments ...

the guitar, the tres
