### 2023-05-10 #5395 @mrocamora
> Here we continue the analysis guitar and tres and calculate the asynchrony between them.

the analysis with the guitar and the tres
