### 2023-05-10 #5395 @mrocamora
> shows that bass is ahead (ie. -19 ms)

i.e.
