### 2023-05-10 #5395 @mrocamora
> we will plot the estimated frequency curves (Figure 6).

we plot the estimated ...
