### 2023-05-10 #5395 @mrocamora
Regarding the review checklist, I could not find community guidelines.

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
