### 2023-05-10 #5395 @mrocamora
Regarding the review checklist, I could not find a comparison to  other commonly-used packages. This is probably because there are no other packages of this kind.

> State of the field: Do the authors describe how this software compares to other commonly-used packages?
