### 2023-05-11 #5395 @mrocamora
@tuomaseerola I went through the paper, the repo, and the web and checked the examples' functionality. It is an excellent contribution with relevance in several research scenarios. The word count for the paper.md file is 4834 (according to the editorial bot), surpassing the word limit, so we should check with @faroit on how to proceed. I have noted some typos and minor errors that are easy to fix and are listed above. Besides, some of the points in the review checklist are missing. Community guidelines should be added. As for the rest, automated tests and comparison to other packages, we should confirm with @faroit if they are mandatory.
