### 2023-05-13 2d4660a
Update paper.Rmd
Fixed a typo
