### 2023-05-13 e34195d
Removed excess details for DOIs
removed http://dx.doi.org/ from all DOIs
