+++ paper/paper.bib
@@ -42,13 +42,12 @@
 	author = {Eerola, T. and Jakubowski, K. and Moran, N. and Keller, P. and Clayton, M.},
 	date-added = {2023-01-17 08:33:33 +0000},
 	date-modified = {2023-01-17 08:33:33 +0000},
-	doi = {http://dx.doi.org/10.1098/rsos.171520},
+	doi = {10.1098/rsos.171520},
 	journal = {Royal Society Open Science},
 	pages = {171520},
 	title = {Shared Periodic Performer Movements Coordinate Interactions in Duo Improvisations},
 	volume = {5},
-	year = {2018},
-	bdsk-url-1 = {http://dx.doi.org/10.1098/rsos.171520}}
+	year = {2018}}
 
 @article{goebl2010investigations,
 	author = {Goebl, Werner and Flossmann, Sebastian and Widmer, Gerhard},
@@ -175,8 +174,7 @@
 	title = {{IEMP} {C}uban {S}on and {S}alsa},
 	url = {osf.io/sfxa2},
 	year = {2022},
-	bdsk-url-1 = {osf.io/sfxa2},
-	bdsk-url-2 = {https://doi.org/10.17605/OSF.IO/SFXA2}}
+	bdsk-url-1 = {osf.io/sfxa2}}
 
 @misc{Clayton_Leante_Tarsitani_2021,
 	author = {Clayton, Martin and Leante, Laura and Tarsitani, Simone},
@@ -187,8 +185,7 @@
 	title = {{IEMP} {N}orth {I}ndian {R}aga},
 	url = {osf.io/ks325},
 	year = {2021},
-	bdsk-url-1 = {osf.io/ks325},
-	bdsk-url-2 = {https://doi.org/10.17605/OSF.IO/KS325}}
+	bdsk-url-1 = {osf.io/ks325}}
 
 @article{wing2014optimal,
 	author = {Wing, Alan M and Endo, Satoshi and Bradbury, Adrian and Vorberg, Dirk},
@@ -230,29 +227,27 @@
 	author = {Clayton, M. and Jakubowski, K. and Eerola, T.},
 	date-added = {2023-01-17 08:33:33 +0000},
 	date-modified = {2023-01-17 08:33:33 +0000},
-	doi = {https://doi.org/10.1177/1029864919844809},
+	doi = {10.1177/1029864919844809},
 	issue = {3},
 	journal = {Musicae Scientiae},
 	keywords = {entrainment},
 	pages = {304-331},
 	title = {Interpersonal entrainment in {I}ndian instrumental music performance: Synchronization and movement coordination relate to tempo, dynamics, metrical and cadential structure},
 	volume = {23},
-	year = {2019},
-	bdsk-url-1 = {https://doi.org/10.1177/1029864919844809}}
+	year = {2019}}
 
 @article{Clayton2020emr,
 	author = {Clayton, M and Tarsitani, S and Jankowsky, R and Jure L and Leante, L and Polak, R and Poole, A and Rocamora, M and Alborno, P and Camurri, A and Eerola, T and Jacoby, N and Jakubowski, K},
 	date-added = {2023-01-17 08:33:33 +0000},
 	date-modified = {2023-01-17 08:33:33 +0000},
-	doi = {http://dx.doi.org/10.18061/emr.v16i1.7555},
+	doi = {10.18061/emr.v16i1.7555},
 	journal = {Empirical Musicology Review},
 	number = {1},
 	pages = {65-84},
 	title = {The Interpersonal Entrainment in Music Performance Data Collection},
-	url = {http://dx.doi.org/10.18061/emr.v16i1.7555},
+	url = {10.18061/emr.v16i1.7555},
[…]
