+++ paper/paper.Rmd
@@ -53,13 +53,13 @@ through note onsets. Data is ingested in the form of CSV files with individual i
 
 ## 1.2 Onset datasets
 
-`onsetsync` package comes with example performances from two traditions, Cuban Son and Salsa performances [@Poole_Tarsitani_Clayton_2022] and North Hindustani performances [@Clayton_Leante_Tarsitani_2021]. The examples are taken from the Interpersonal Entrainment in Music Performance (IEMP) project collection that has compiled small ensemble performance data from six traditions (North Indian ragas, Malian jembe, Tunisian stambeli, Uruguayan candombe, European string quartet, and Cuban son and salsa). These collection contain tens of hours annotated and verified onsets and are available from Open Science Framework (OSF). There are also noteworthy datasets containing audio and annotated contents of music relevant to assessing synchrony such as _Carnatic Music Rhythm Dataset_ [@srinivasamurthy2015particle], _Hindustani Music Rhythm Dataset_ [@srinivasamurthy2016generalized], _Arab-Andalusian (Maghreb) dataset_ [@sordo2014creating], _Scandinavian Fiddle Tunes with Accompanying Foot-Tapping_ [@lordelo2020adversarial] and _MAESTRO (MIDI and Audio Edited for Synchronous Tracks and Organisation)_ dataset which is an extensive collection (200+ hours) of professional piano performances [@hawthorne2018enabling]. Some datasets such as Chopin performances compiled by [@goebl2010investigations] facilitate study of the synchrony between the two hands of the pianist. 
+`onsetsync` package comes with example performances from two traditions, Cuban Son and Salsa performances [@Poole_Tarsitani_Clayton_2022] and North Hindustani performances [@Clayton_Leante_Tarsitani_2021]. The examples are taken from the Interpersonal Entrainment in Music Performance (IEMP) project collection that has compiled small ensemble performance data from six traditions (North Indian ragas, Malian jembe, Tunisian stambeli, Uruguayan candombe, European string quartet, and Cuban son and salsa). These collections contain tens of hours annotated and verified onsets and are available from Open Science Framework (OSF). There are also noteworthy datasets containing audio and annotated contents of music relevant to assessing synchrony such as _Carnatic Music Rhythm Dataset_ [@srinivasamurthy2015particle], _Hindustani Music Rhythm Dataset_ [@srinivasamurthy2016generalized], _Arab-Andalusian (Maghreb) dataset_ [@sordo2014creating], _Scandinavian Fiddle Tunes with Accompanying Foot-Tapping_ [@lordelo2020adversarial] and _MAESTRO (MIDI and Audio Edited for Synchronous Tracks and Organisation)_ dataset which is an extensive collection (200+ hours) of professional piano performances [@hawthorne2018enabling]. Some datasets such as Chopin performances compiled by @goebl2010investigations facilitate study of the synchrony between the two hands of the pianist. 
 
 ## 1.3 Terminology and measures of synchrony
 
 Much of the research on synchrony in music has focused on
 _sensorimotor synchronization_ (SMS) or just synchronization, a
-processes occurring within 100-2000 ms timescales, which has commonly
+process occurring within 100-2000 ms timescales, which has commonly
 been studied with tapping experiments [@repp2008sensorimotor, @repp2013sensorimotor] but also through analyses of music corpora, from string quartets [@wing2014optimal] to drum ensembles [@polak2016both] and to
 Indian instrumental music performances [@clayton_et_al_2018b]. Synchronisation analyses can also be applied to field observations
 such as the study of how walking drumming groups attempt to
