### 2023-05-13 #5395 @tuomaseerola
> Eq. 1: sum symbol should say `i`

Is this about Eq. 2 which has sum symbol?
