### 2023-05-13 tuomaseerola/onsetsync#1 @tuomaseerola
All DOIs are now fixed
