### 2023-05-13 #5395 @faroit
@mrocamora @pmcharrison thank you already for your very detailed reviews. I value that you also went deeper into the notation of the paper. However, to free this issue up from too many comments, I suggest to do future discussion about notation in distinct issues in the submission repository.

E.g. we encourage reviewers to file issues against the submitted repository’s issue tracker. You can include in your review links to any new issues that you the reviewer believe to be impeding the acceptance of the repository.
