### 2023-05-13 #5395 @faroit
@mrocamora

>  The word count for the paper.md file is 4834 (according to the editorial bot), surpassing the word limit.

even if most of the word count is driven by the tables, i would agree that the paper is too long and should be shorten by at least 1-2 pages.

please open an issue to track this.

> Community guidelines should be added.

Yes, please open an issue.

> As for the rest, automated tests and comparison to other packages, we should confirm with @faroit if they are mandatory.

automated test are preferred but not absolutely mandatory if step-by-step instructions are provided. In this case however, I would suggest to add a minimal set of automation (e.g. package installation, running)

I think a comparison (and references) to other packages are highly suggested, at least when it comes to general onset detection and evaluation.
