### 2023-05-14 tuomaseerola/onsetsync#3 @mrocamora
Word limit - JOSS paper
@tuomaseerola the word count for the paper.md file is 4834 (according to the editorial bot), surpassing the word limit.
Editor (@faroit) suggests that even if most of the word count is driven by the tables, the paper is too long and should be shorten by at least 1-2 pages.
