### 2023-05-14 tuomaseerola/onsetsync#4 @mrocamora
Community guidelines - JOSS paper
@tuomaseerola community guidelines should be added.

Clear guidelines for third parties wishing to:
1) Contribute to the software
2) Report issues or problems with the software
3) Seek support
