### 2023-05-14 tuomaseerola/onsetsync#5 @mrocamora
Automated tests - JOSS review
@tuomaseerola  According to editor (@faroit) automated test are preferred but not absolutely mandatory if step-by-step instructions are provided. In this case however, he suggests to add a minimal set of automation (e.g. package installation, running).
