### 2023-05-14 #5395 @mrocamora
@faroit @tuomaseerola @pmcharrison I've added an issue regarding word limit

https://github.com/tuomaseerola/onsetsync/issues/3#issue-1709098619
