### 2023-05-14 #5395 @mrocamora
@faroit @tuomaseerola @pmcharrison I've added and issue about the community guidelines

https://github.com/tuomaseerola/onsetsync/issues/4#issue-1709099132
