### 2023-05-14 #5395 @mrocamora
@faroit @tuomaseerola @pmcharrison I've added and issue to discuss the automation (test, install, etc.)

https://github.com/tuomaseerola/onsetsync/issues/5#issue-1709099745
