### 2023-05-28 tuomaseerola/onsetsync#4 @tuomaseerola
Thanks, `CONTRIBUTING.md` document has now been added to the repository containing these topics and a link has been created to `README.md` file.
