+++ CONTRIBUTING.md
@@ -1,16 +1,16 @@
 # How-to contribute
 
-Thanks for thinking about contributing to the onsetsync R package. This is an open source project and improvements and additions are welcome.
+Thanks for thinking about contributing to the [`onsetsync`](https://github.com/tuomaseerola/onsetsync) [R package](https://www.r-project.org). This is an open source project and improvements and additions are welcome.
 
-1. Contribute to the software
+## Contribute to the software
 
 You can contribute to different ways to the project. There is a tutorial on contributing to [Open Source projects](https://opensource.guide/how-to-contribute/) and using [Github with R](https://happygitwithr.com/rstudio-git-github.html). There is also a useful [Code of Conduct](https://github.com/github/docs/blob/main/CODE_OF_CONDUCT.md) for Github which keeps the community relations intact.
 
-2. Report issues or problems with the software
+## Report issues or problems with the software
 
 If you have questions, or you want to raise an issue or problem, you can post it as an [issue on Github](https://github.com/tuomaseerola/onsetsync/issues/new). While we cannot promise user support, we will try to address it, as questions will improve the documentation and code. You can also suggest new features or report issues in the website (which has been created with [pkgdown](http://pkgdown.r-lib.org/)).
 
-3. Seek support
+## Seek support
 
 If you want to ask the question in private, feel free to contact the developers by email. If you want to learn more about the concepts behind synchrony, do browse the core literature mentioned in the documentation and check the articles in the [Documentation](https://tuomaseerola.github.io/onsetsync/articles/onsetsync.html) and the [paper](https://github.com/tuomaseerola/onsetsync/blob/master/paper/paper.md).
 
