### 2023-05-28 cc49944
Added link to CONTRIBUTING.md
