### 2023-05-28 tuomaseerola/onsetsync#3 @tuomaseerola
Of course, this silly of us not following the upper limit. We have now eliminated half of the content from the paper (it is now 2508 words) by taking several sections and multiple graphs as additional documentation articles:

- [Advanced topics](https://tuomaseerola.github.io/onsetsync/articles/advanced_topics.html) (which contains 3 sub-topics from the paper)
- [Analysis of periodicity](https://tuomaseerola.github.io/onsetsync/articles/analysisis_of_periodicity.html)

We also streamlined the text to be shorter.
