### 2023-05-28 tuomaseerola/onsetsync#5 @tuomaseerola
Indeed. The first 3 code block cover this (#1 is install, #2 is loading data, #3 is visualising the onsets). Moreover, the [`README.md`](https://github.com/tuomaseerola/onsetsync/tree/R1) does this also. Unit-tests have also been added to the package to make sure development does not break calculations, see: https://github.com/tuomaseerola/onsetsync/tree/R1/tests/testthat
