### 2023-05-28 tuomaseerola/onsetsync#5 @faroit
@tuomaseerola I am not that familiar with the R ecosystem but it looks like there are best practices for R when it comes to automated testing https://github.com/bradleyboehmke/unit-testing-r

Moreover, the readme has instructions how to run the software but I fail to see where actual tests are called. Can you let us know where this is happening, please?
