### 2023-05-28 #5395 @tuomaseerola
> Typos in this sentence:
>
> > These collection contain tens of hours annotated and verified

fixed
