### 2023-05-28 #5395 @tuomaseerola
> Wrong citation format here:

>A number of different measures of synchrony have been proposed. We follow the definitions
> 54 from (Clayton et al., 2020, p. 161)

fixed (and Rasch references altered as well).
