### 2023-05-28 #5395 @tuomaseerola
> Eq. 1: sum symbol should say `i`

i=1 added to the sum of the Eq. 2
