### 2023-05-28 #5395 @tuomaseerola
> Eq. 3: why not use `d_i` abbreviation here?

replaced the clunky notation with `d_i`
