### 2023-05-28 #5395 @tuomaseerola
> Eq. 3: `T` has not been defined yet

Switched to `n` to be consistent and defined it as "where `n` is the number of onset time differences. "
