### 2023-05-28 #5395 @tuomaseerola
> Eq. 4: `i=i` should be `i=1`, same in Eq. 5

fixed
