### 2023-05-28 #5395 @tuomaseerola
> Eq. 5: isn't this the same as d bar, used in Eq. 2? If so, use the same symbol for both?

fixed
