### 2023-05-28 #5395 @tuomaseerola
> Eq. 6: shouldn't the intro to this equation say onset time `v` instead of onset time `w`?

Yes, fixed, and all $\hat{w}$ symbols are now $\bar{w}$ symbols for consistency.
