### 2023-05-28 #5395 @tuomaseerola
> Eq. 8: `i=i` should be `i=1`.
>
> I'm also not sure what the subscript `r` means here.

fixed `i=1`, removed `r` as subscript, and switched $\hat{s}$ to $\bar{s}$ to refer to mean.
