### 2023-05-28 #5395 @tuomaseerola
> Table 2: it seems unfortunate that SD is used as a column name, because most readers will think this denotes standard ?> deviation. Probably it's too late to change now though

Switching from SD to Beat would make this variable clearer. It requires extensive revision, so will first estimate the amount of work needed.
