### 2023-05-28 #5395 @tuomaseerola
> Typo:
>
> > As a broad overview, we can use the function plot_by_beat visualise the asynchrony relative to

fixed (sentence has been eliminated)
