### 2023-05-28 #5395 @tuomaseerola
> 'granger' should be capitalised

fixed
