### 2023-05-28 #5395 @tuomaseerola
> Regarding this checklist point,
>
> > Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?
>
> I could not see any automated tests, but maybe I am missing them?

The first three code blocks will do this and the [Documentation Quick Guide](https://tuomaseerola.github.io/onsetsync/articles/onsetsync.html) starts with this and the primary vignette ('onsetsync') will also have the basic steps coded in with build-in data.
