### 2023-05-28 #5395 @tuomaseerola
> Regarding this checklist point,
>
> > Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
>
> I could not see any, but maybe I am missing them?

These have now been added as a separate file `CONTRIBUTING.md` and linked from the landing page and documentation.
