### 2023-05-28 #5395 @tuomaseerola
> Regarding this checklist point,
>
> > State of the field: Do the authors describe how this software compares to other commonly-used packages?
>
> I could not see any comparison, but maybe I missed it.

There is a Python library [`SyncPy`](https://github.com/syncpy/SyncPy) that could provide some overlapping functionality but since it is not commonly used and it has not been updated since 2017, flagging this up felt unnecessary.
