### 2023-05-28 #5395 @tuomaseerola
> There is a typo in this reference:
>
> > Müller m, Wiering f, Editors.
>
> and this one:
>
> > A generalized bayesian

fixed both (and some other minor bibtex quirks)
