### 2023-05-28 #5395 @tuomaseerola
> > Here we first define the onset time differences, 𝑑, ...
>
> Then, in the definition of the precision measures, the term "asynchronies" (or "asynchrony values") is used for d, instead of "onset time differences". So I suggest: "Here we first define the onset time differences (or asynchronies), 𝑑, ..."

thanks, this has been adopted. The terminology needs tightening up.
