### 2023-05-28 #5395 @tuomaseerola
> Eq. 3: According to (Clayton et al., 2020, p. 161), its "the RMS of the pair-wise measurements" of asynchronies. So I guess it should be I1,i−I2,i or just di. Besides, what T stands for should be clarified.

Both issues fixed (`d_i` and replace `T` with `n`.
