### 2023-05-28 #5395 @tuomaseerola
> Eq. 4 and 5: Summation starts at i=1 Eq.4: I suggest using |s¯| instead of |s^| to be consistent with the use of d¯ in Eq. 2.

fixed
