### 2023-05-28 #5395 @tuomaseerola
> Eq. 5: Seems that d¯ should be used instead of sp^ (as in Eq. 2). I don't see where the p comes from (pairwise? if this is the case, note that it is not used in Eq. 2).

fixed
