### 2023-05-28 #5395 @tuomaseerola
> > Mean relative asynchrony : mean position of an instrument’s onsets relative to average position (ω^) of the group, ...
>
> The position of an instrument's onsets was previously notated as I (I1,i and I2,i in Eq. 1). I suggest using a consistent notation for onsets. Besides, I also suggest using w¯ instead of w^, as previously done in Eq. 2.

These are great suggestions, these equations had inconsistencies in them and now $\bar{I}$ instead of $\hat{w}$.
