### 2023-05-28 #5395 @tuomaseerola
> Eq. 8: I suggest using sr¯ instead of sr^.

fixed
