### 2023-05-28 #5395 @tuomaseerola
> > (e.g. Librosa, Essentia, or MIR toolbox for Matlab, or Sonic Visualiser using well-known onset detection algorithms.
>
> There is an extra "or" and the closing parenthesis is missing.

both issues fixed, thank you.
