### 2023-05-28 #5395 @tuomaseerola
> > see (Schlüter & Böck, 2014) and (Böck et al., 2016), and for a full workflow of how to combine onset detection and annotation of the musical information, see (Clayton, Tarsitani, et al., 2021).
>
> Wrong citation format.

fixed
