### 2023-05-28 #5395 @tuomaseerola
> > (with precision of 6 digits, ...
>
> Closing parenthesis is missing.

fixed
