### 2023-05-28 #5395 @tuomaseerola
> Information about cycles and beat subdivisions is generally based on manual annotation, and originates separately from the automatically extracted onsets; the two are combined to produce tables with onsets assigned to beat positions.

Good, a clarification has been added:
> This explains why the first row of the table has no onsets but it contains annotation information (Cycle 1 and beat sub-division 1)
