### 2023-05-28 #5395 @tuomaseerola
> Sec 1.6 Onset summaries
>
> Note that you need to load `library(dplyr)` to run the example code, as it internally uses the `ungroup` function. This is clearly specified in the Usage section in the repository but not in the paper. I suggest adding a Usage section to the paper. I am aware of the "Analysis Example" vignette, which has all the code needed to run the examples. However, the code to load the libraries is small enough to be included in the paper and could allow someone reading the paper to follow it by copy-pasting the code in R.

We agree with this and adding the explicit call to the library is short price to pay for the easy of use.
