### 2023-05-28 #5395 @tuomaseerola
> Table 3. : It seems that SD here is different from that of Table 2.

This table has now been omitted to save space but the point this stands; SD is the common abbreviation of standard deviation and unfortunately seems to create confusion. We will review during the revision whether switching from the sub-division (SD) to something else is the best solution.
