### 2023-05-28 #5395 @tuomaseerola
> > Overall this suggests that these two instruments ...  although guitar is tightly aligned ... are not evident in tres ...
>
> the guitar, the tres

fixed
