### 2023-05-28 #5395 @tuomaseerola
> > Here we continue the analysis guitar and tres and calculate the asynchrony between them.
>
> the analysis with the guitar and the tres

This section is no longer in the article but has been fixed in the Documentation (articles)
