### 2023-05-28 #5395 @tuomaseerola
> > We constrain the frequency a range of
>
> Typo: to a range

This section is no longer in the article but has been fixed in the Documentation (articles)
