### 2023-05-28 #5395 @tuomaseerola
> > we will plot the estimated frequency curves (Figure 6).
>
> we plot the estimated ...

This section is no longer in the article but has been fixed in the Documentation (articles)
