### 2023-05-28 #5395 @tuomaseerola
> Regarding the review checklist, I could not find automated tests.
>
> > Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

This has been addressed in the documentation and also the three first code blocks in the manuscript function as a test of the functionality of the software using the build-in data.
