### 2023-05-28 #5395 @tuomaseerola
> Regarding the review checklist, I could not find a comparison to other commonly-used packages. This is probably because there are no other packages of this kind.
>
> > State of the field: Do the authors describe how this software compares to other commonly-used packages?

True, nothing too directly similar. There is a Python library `SyncPy` that could provide some overlapping functionality but since it is not commonly used and it has not been updated since 2017, flagging it up felt unnecessary.
