### 2023-05-28 #5395 @tuomaseerola
> @tuomaseerola I went through the paper, the repo, and the web and checked the examples' functionality. It is an excellent contribution with relevance in several research scenarios. The word count for the paper.md file is 4834 (according to the editorial bot), surpassing the word limit, so we should check with @faroit on how to proceed. I have noted some typos and minor errors that are easy to fix and are listed above. Besides, some of the points in the review checklist are missing. Community guidelines should be added. As for the rest, automated tests and comparison to other packages, we should confirm with @faroit if they are mandatory.

Thank you for the diligent and careful reading of the manuscript and you clearly explored the repo and tested the functionality of the commands. Really constructive comments and we have incorporated them all as best as we could. We hope radical shortening and removing several sections into online Documentation will make the issue about the length less of an issue for the journal.
