### 2023-05-28 tuomaseerola/onsetsync#5 @tuomaseerola
@faroit This is valuable idea and while the devtools::check will run all code automatically and report conflicts and errors, an explicit unit-testing would make the actual calculation more reliable against version changes and other issues. We have now added a standard 'testthat' workflow for unit testing for (1) reading data, (2) asynchrony calculation, (3) plotting, and (3) periodicity. These now report positively:
> devtools::test()
ℹ Testing onsetsync
✔ | F W S  OK | Context
✔ |         2 | 01-read-data
✔ |         2 | 02-asynchrony
✔ |         1 | 03-figures
✔ |         1 | 04-periodicity
══ Results ══════════════
Duration: 0.3 s

[ FAIL 0 | WARN 0 | SKIP 0 | PASS 6 ]

The unit tests are now included in the new release (code is in the `\tests` subfolder). https://github.com/tuomaseerola/onsetsync/tree/R1/tests/testthat
