+++ paper/paper.bib
@@ -1,7 +1,7 @@
 %% This BibTeX bibliography file was created using BibDesk.
 %% https://bibdesk.sourceforge.io/
 
-%% Created for Tuomas Eerola at 2023-01-17 08:35:09 +0000 
+%% Created for Tuomas Eerola at 2023-05-28 19:25:24 +0100 
 
 
 %% Saved with string encoding Unicode (UTF-8) 
@@ -47,7 +47,8 @@
 	pages = {171520},
 	title = {Shared Periodic Performer Movements Coordinate Interactions in Duo Improvisations},
 	volume = {5},
-	year = {2018}}
+	year = {2018},
+	bdsk-url-1 = {https://doi.org/10.1098/rsos.171520}}
 
 @article{goebl2010investigations,
 	author = {Goebl, Werner and Flossmann, Sebastian and Widmer, Gerhard},
@@ -106,11 +107,11 @@
 @article{wallot2018analyzing,
 	author = {Wallot, Sebastian and Leonardi, Giuseppe},
 	date-added = {2023-01-17 08:33:33 +0000},
-	date-modified = {2023-01-17 08:33:33 +0000},
+	date-modified = {2023-05-28 19:25:24 +0100},
 	journal = {Frontiers in Psychology},
 	pages = {2232},
 	publisher = {Frontiers Media SA},
-	title = {Analyzing multivariate dynamics using cross-recurrence quantification analysis ({CRQA}), diagonal-cross-recurrence profiles (dcrp), and multidimensional recurrence quantification analysis ({MDRQA})--a tutorial in {R}},
+	title = {Analyzing multivariate dynamics using cross-recurrence quantification analysis ({CRQA}), diagonal-cross-recurrence profiles (dcrp), and multidimensional recurrence quantification analysis ({MDRQA})-- {A} tutorial in {R}},
 	volume = {9},
 	year = {2018}}
 
@@ -150,19 +151,21 @@
 	author = {Srinivasamurthy, Ajay and Holzapfel, Andre and Cemgil, Ali Taylan and Serra, Xavier},
 	booktitle = {2016 IEEE International Conference on Acoustics, Speech and Signal Processing (ICASSP)},
 	date-added = {2023-01-17 08:33:33 +0000},
-	date-modified = {2023-01-17 08:33:33 +0000},
+	date-modified = {2023-05-28 19:24:06 +0100},
 	organization = {IEEE},
 	pages = {76--80},
-	title = {A generalized bayesian model for tracking long metrical cycles in acoustic music signals},
+	title = {A generalized {B}ayesian model for tracking long metrical cycles in acoustic music signals},
 	year = {2016}}
 
 @inproceedings{srinivasamurthy2015particle,
+	address = {M{\'a}laga, Spain},
 	author = {Srinivasamurthy, Ajay and Holzapfel, Andre and Cemgil, Ali Taylan and Serra, Xavier},
-	booktitle = {M{\"u}ller M, Wiering F, editors. ISMIR 2015. 16th International Society for Music Information Retrieval Conference; 2015 Oct 26-30; M{\'a}laga, Spain. Canada: ISMIR; 2015.},
+	booktitle = {Proceedings of the 16th International Society for Music Information Retrieval Conference; 2015 Oct 26-30},
 	date-added = {2023-01-17 08:33:33 +0000},
-	date-modified = {2023-01-17 08:33:33 +0000},
+	date-modified = {2023-05-28 19:23:24 +0100},
+	editor = {M{\"u}ller M. and Wiering F.},
 	organization = {International Society for Music Information Retrieval (ISMIR)},
[…]
