+++ paper/paper.Rmd
@@ -32,10 +32,10 @@ journal: JOSS
 
 Music performance relies on tight yet flexible timing coordination
 between performers. An important aspect of this interpersonal
-entrainment is the synchronisation of musical events between performers,
+entrainment is the synchronization of musical events between performers,
 and this can be analysed from the note onsets obtained from recorded
 performances. Analysis of the asynchronies between onsets gives an
-insight into synchronisation between performers, and the extent to which
+insight into synchronization between performers, and the extent to which
 it is accurate (how close together the events are in time) and precise
 (how stable the relationship is). The synchrony between performers is
 influenced by various factors such as the genre of music, performer
@@ -58,17 +58,14 @@ through note onsets. Data is ingested in the form of CSV files with individual i
 Much of the research on synchrony in music has focused on
 _sensorimotor synchronization_ (SMS) or just synchronization, a
 process occurring within 100-2000 ms timescales, which has commonly
-been studied with tapping experiments [@repp2008sensorimotor, @repp2013sensorimotor] but also through analyses of music corpora, from string quartets [@wing2014optimal] to drum ensembles [@polak2016both] and to
-Indian instrumental music performances [@clayton_et_al_2018b]. Synchronisation analyses can also be applied to field observations
-such as the study of how walking drumming groups attempt to
-resist synchronisation to the beats of other drum groups during their
-walk across the city during a festival [@lucas2011inter].
+been studied with tapping experiments [@repp2008sensorimotor; @repp2013sensorimotor] but also through analyses of music corpora, from string quartets [@wing2014optimal] to drum ensembles [@polak2016both], and to
+Indian instrumental music performances [@clayton_et_al_2018b].
 
 A number of different measures of synchrony have been proposed. We
-follow the definitions from [@clayton2020, p. 161] who adopt and
-develop the measures defined by [@rasch1988timing, @rasch1979synchronization]:
+follow the definitions from @clayton2020 who adopt and
+develop the measures defined by Rasch [-@rasch1988timing; -@rasch1979synchronization]:
 
-Here we first define the onset time differences, $d$, where $I_{1, i}$ and $I_{2, i}$ refer to an onset at time point $i$ for instrument 1 and 2:
+Here we first define the onset time differences (or asynchronies), $d$, where $I_{1, i}$ and $I_{2, i}$ refer to an onset at time point $i$ for instrument 1 and 2:
 
 \begin{equation}
 d_{i} = I_{1, i} - I_{2, i}
@@ -80,55 +77,53 @@ d_{i} = I_{1, i} - I_{2, i}
   asynchronies of any pair of instruments, which is
 
 \begin{equation}
-Sd = \sqrt\frac{\sum{(d_i-\bar{d})^2}}{n-1}
+Sd = \sqrt\frac{\sum_{i=1}{(d_i-\bar{d})^2}}{n-1}
 \end{equation}
 
 *Groupwise asynchronization*: Root mean square (RMS) of the
   pairwise asynchronizations calculated as
 
 \begin{equation}
-RMS = \sqrt{\Sigma_{i=1}^{T}{\Big(\frac{d_{1,i} -d_{2,i}}{T}\Big)^2}}
+RMS = \sqrt{\Sigma_{i=1}^{n}{\Big(\frac{d_{i}}{n}\Big)^2}}
 \end{equation}
+
+where $n$ is the number of onset time differences. 
 
 *Mean absolute asynchrony*: Mean of all unsigned asynchrony
   values which is 
 
 \begin{equation}
-\hat{|s|} = \frac{1}{n} \sum_{i=i}^{n} |d_{i}|
[…]
