### 2023-05-29 #5395 @pmcharrison
Thanks @tuomaseerola -- is your next draft complete, then? Shall we regenerate the paper and have another read?
