+++ R/plot_by_dataset.R
@@ -24,25 +24,11 @@ plot_by_dataset <-
            colour = 'lightblue',
            colorpalette = 'Set2') {
     # T. Eerola, Durham University, IEMP project
-    # 23/1/2018
-    # needs work
-    
-    # for testing
-    # df<-asere
-    # instr<-c('Bass','Tres','Bell')
-    # beat<-'SD'
-    # virtual<-'Virtual.SD'
-    # pcols<-2
-    # box<-TRUE
-    # griddeviations<-FALSE
-    #print(data)
-    #print(asynchronies)
     asynch <- n <- NULL
     
     DF <- dplyr::select(df, asynchronies, data)
     colnames(DF) <- c('asynch', 'data')
-#    print(head(DF))
-    
+
     g1 <-
       ggplot2::ggplot(DF, ggplot2::aes(asynch, fill = data), colour = 'black') +
       ggplot2::geom_density(alpha = 1, show.legend = FALSE) +
