### 2023-06-30 #5395 @faroit
> Sure, @faroit, I thought I did this on May 29 already, so all revisions (including the corrections requested to the paper, and unit tests, and additional articles for documentation that were discarded from the paper) have been completed the revisions (R1 branch) branch, although I have not merged these with the main branch as I assumed you would be able to see the branch. Should I merge R1 branch with the main branch?

No, that's fine. I just wasn't just if this meant that the reviews were addresses in there. Merging can be done at the end of the review process before a new tag will be issued.
