### 2023-07-07 #5395 @pmcharrison
> > Eq. 3: `T` has not been defined yet
>
> Switched to `n` to be consistent and defined it as "where `n` is the number of onset time differences. "

I can't see this change in the latest copy?
