### 2023-07-21 #5395 @pmcharrison
Thanks @tuomaseerola, @faroit, I apologise for not keeping track of the conversation above about the location of the edits. @faroit, you ran generate PDF recently, but this will have generated the PDF for the wrong branch, no? It is hard for us to review the revisions without this. @tuomaseerola one way to facilitate the review would be if you create a PR comparing master with R1, then we can see the diff.
