### 2023-07-28 #5395 @faroit
@tuomaseerola @pmcharrison i think the easiest would be to merge into main now for the pdf and track the changes in https://github.com/tuomaseerola/onsetsync/compare/master...R1
