### 2023-07-28 #5395 @mrocamora
Hi @faroit Yes, the community guidelines and the automated tests have been added. Should I close the two issues?
