### 2023-07-28 #5395 @mrocamora
I am with @pmcharrison in that we cannot see the changes in the generated pdf.
Once this is solved (@pmcharrison  suggested a PR), I can check the minor typos and corrections.
