### 2023-08-31 #5395 @pmcharrison
Hi all, I'm very sorry that this has taken a while to sort out. I'm really happy with the revised paper and think it's good to go. I think it's going to be a really useful resource and inspiration for people working on this topic.

A minor note is that I think Figure 2 could be improved slightly:

The simplest improvement would be to reduce the % label size so that the labels don't overlap. That would make the plot look more production ready.

If I were to be more picky, I would say that the label representation is not the best way to represent the data. The point of making plots is that one should not have to resort to listing numbers like we do in tables. I think it would be preferable if these mean offsets were presented instead by a marginal line plot. I appreciate that this is a bit more complicated to achieve in R, but I believe you could achieve it after a bit of work via `grid.arrange` (see this example) https://stackoverflow.com/a/17649351.

All in all, though, I don't think this consideration is that important, and you should feel absolutely free to leave it as is.
