### 2023-09-06 #5395 @tuomaseerola
I agree. The minimal edit is to adjust the size of the labels and I also that it is not optimal to have visualisation and tabular information combined, but this was user request initially. I'll probably do a new switch that plots these as marginal lines and leave the unlabelled as the default, and retain the percentage labelled (with size adjustment) as an option as well.
