### 2023-09-06 ee2d39f
Updated the text labelling (smaller, explicit sign)
