### 2023-09-06 5fd5cee
Update plot_by_beat.R
changed text label size
