+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: onsetsync
 Title: Analysis and Visualisation of Synchronisation of Music Onset Data
-Version: 0.4.9
+Version: 0.5.0
 Authors@R: person("Tuomas", "Eerola", email = "tuomas.eerola@durham.ac.uk", role = c("aut", "cre"))
 Description: Helps to calculate and visualise onset data. The emphasis is on 
     comparing onsets between performers in the same ensemble, in other words,
