### 2023-09-09 d26174e
Update paper.Rmd
install options changes
