### 2023-09-09 ccc97fc
Update plot_by_beat.R
killed an extra argument in geom_label
