### 2023-09-09 e03b96f
updated visuals on plot_by_beat
