### 2023-09-11 #5395 @mrocamora
Yes, I think this is solved. There is no need to mention SyncPy and there are several examples of tools for onset extraction.
