### 2023-09-11 #5395 @mrocamora
Y confirm that the new version looks great too! @faroit I am happy to go ahead.

BTW please note that lines 74 to 89 of the pdf (the R CMD build output) should be removed.
