### 2023-09-13 #5395 @faroit
@tuomaseerola happy to continue with the last steps now.

> BTW please note that lines 74 to 89 of the pdf (the R CMD build output) should be removed.

can you modify this, please and let me know when the code is up-to-date?
