### 2023-09-14 #5395 @tuomaseerola
> @tuomaseerola happy to continue with the last steps now.
>
> > BTW please note that lines 74 to 89 of the pdf (the R CMD build output) should be removed.
>
> can you modify this, please and let me know when the code is up-to-date?

Yes, @faroit, the installation message is now "quiet" and the actual installation refers to the main repo, not to the R1 branch, which means I need to merge R1 with the main once this review process is complete. For changes, see
https://github.com/tuomaseerola/onsetsync/commit/549b9fe44d6ba9184ce65159ff3fed1549925d16
