### 2023-09-23 #5395 @faroit
@pmcharrison i understand that this would be useful but I feel as this time this would be out of scope for the JOSS submission as it's a about to get accepted soon. Cant you just create a GitHub issue?
