### 2023-09-25 tuomaseerola/onsetsync#12 @pmcharrison
Document format of input datasets
@tuomaseerola I had a student try to run some analyses using the package and they struggled to work out what kind of format is required when providing one's own dataset. There is some information on this in the paper but I think it would be helpful to document it in the package itself too, ideally for all functions that ingest such datasets (roxygen has functionality for inheriting documentation so you don't have to copy and paste each time). Ideally this documentation would explain which columns are mandatory and which are optional. I also found the definition of Isochronous.SD.Time a bit confusing and wondered whether it could be explained a bit clearer, e.g. through the use of an example?
