### 2023-09-25 #5395 @pmcharrison
I have filed an issue too now. I do recommend it be addressed before publication though, otherwise anyone who sees the package published and tries to run it on their own data will likely suffer from similar issues.
