### 2023-09-25 #5395 @tuomaseerola
I do recognise the issue myself, the existing examples are way above the "minimal required" information and explaining the required representations (columns of data) would be helpful. I have not thought through whether every function then would have to be checked and signposted for the specific requirements, it probably needs input check routines and documentation for each function, and turns easily into changing the the input to S3 classes but if the minimal change would be to add documentation to the package using a vignette (e.g. "input-representation"), this could be an easy task.
