### 2023-09-26 #5395 @pmcharrison
Hi @tuomaseerola, the vignette would indeed be a nice solution and you should do that if you think it's most efficient. The other (near-?) minimal solution would be to go to `plot_by_beat.R`, and update this line:

```
#' @param df data frame to be processed
```

to specify complete information about the required columns, then in every other function that uses such a dataset you delete that line and instead write

```
@inheritParams plot_by_beat
```
