### 2023-10-05 tuomaseerola/onsetsync#12 @pmcharrison
From @tuomaseerola:

> I do recognise the issue myself, the existing examples are way above the "minimal required" information and explaining the required representations (columns of data) would be helpful. I have not thought through whether every function then would have to be checked and signposted for the specific requirements, it probably needs input check routines and documentation for each function, and turns easily into changing the the input to S3 classes but if the minimal change would be to add documentation to the package using a vignette (e.g. "input-representation"), this could be an easy task.

and my reply:

Hi @tuomaseerola, the vignette would indeed be a nice solution and you should do that if you think it's most efficient. The other (near-?) minimal solution would be to go to `plot_by_beat.R`, and update this line:

```
#' @param df data frame to be processed
```

to specify complete information about the required columns, then in every other function that uses such a dataset you delete that line and instead write

```
#' @inheritParams plot_by_beat
```
