### 2023-10-10 tuomaseerola/onsetsync#12 @tuomaseerola
Hi @pmcharrison. Very cool shortcut from roxygen2, thank you. I opted to document the function inputs explicitly (as the inheritParams only worked for some functions), and all functions now list the required inputs and default parameters in the documentation. I also created a [minimal_representation vignette/article](https://tuomaseerola.github.io/onsetsync/articles/minimal-representation.html). This also led to me to clarify the reference beat naming and calculation, which was initially inherited from the IEMP corpus (the awkward "Isochronous.SD.Time", which will still be evident in some of the documentation when annotated materials are used). If the reference timing is calculated from several instruments, it will be mean onset time and will be labelled as "Mean.Time". If the reference time is inferred from an external reference (Cycle timings from annotations), these are then isochronous time intervals for the cycles and labelled as "Iso.Time".  Documentation and website reflects these updates (this is now 0.5.1 version and the changes are in the R1 branch).
