### 2023-10-11 #5395 @faroit
> @faroit the review is complete on my end, I gave my approval a while ago

@pmcharrison i was referring here to your previous message:

> I have filed an issue too now. I do recommend it be addressed before publication though,

Thus, I thought the closing of the issue was required to complete the review.

In any case, glad to see that we can move on
