+++ README.Rmd
@@ -30,7 +30,7 @@ entering the following commands into R:
 
 ```{r,eval=FALSE}
 if (!require(devtools)) install.packages("devtools")
-devtools::install_github("tuomaseerola/onsetsync@R1",force = TRUE)
+devtools::install_github("tuomaseerola/onsetsync",force = TRUE)
 ```
 
 ## Usage
