+++ paper/paper.md
@@ -127,7 +127,7 @@ These methods are based on asynchronies between onsets occurring at the same bea
 
 
 
-`onsetsync` is available at Github and can be loaded and installed using the code below. The current version is 0.5.0.
+`onsetsync` is available on GitHub and can be loaded and installed using the code below. The current version is 0.5.0.
 
 
 ```r
@@ -136,7 +136,7 @@ devtools::install_github("tuomaseerola/onsetsync")
 library(onsetsync)
 ```
 
-To carry out such comparisons, `onsetsync` has several auxiliary functions. There are functions that help assessing relative timings, to associate onsets with beat subdivisions and cycles in music structure, and to select and compare the instruments. The library has seven categories of functions summarised in Table 1.
+To carry out such comparisons, `onsetsync` has several auxiliary functions. There are functions that help assess relative timings, to associate onsets with beat subdivisions and cycles in music structure, and to select and compare the instruments. The library has seven categories of functions summarised in Table 1.
 
 | Category      | Examples                                             |
 |---------------|------------------------------------------------------|
@@ -150,13 +150,13 @@ To carry out such comparisons, `onsetsync` has several auxiliary functions. Ther
 
 Table: Function categories and example functions or datasets.
 
-In the next sections we will demonstrate the analysis processes related to asynchrony between several performers. It is also worth pointing out that the library is not dedicated to extraction of onsets from audio as that can be done in other packages (e.g. [Librosa](https://librosa.org), [Essentia](https://essentia.upf.edu), [MIR toolbox for Matlab](https://www.jyu.fi/hytk/fi/laitokset/mutku/en/research/materials/mirtoolbox), or [Sonic Visualiser](https://www.sonicvisualiser.org) using well-known onset detection algorithms). Here we take it as granted that the onsets have been extracted already, someone has carried out a quality check, assigned relevant onsets to beat positions and saved them in csv files. For solid overviews on computational onset extraction from audio, see @schluter2014improved or @bock2016madmom, and for a full workflow of how to combine onset detection and annotation of the musical information, see @Clayton2020emr.
+In the next sections, we will demonstrate the analysis processes related to asynchrony between several performers. It is also worth pointing out that the library is not dedicated to extraction of onsets from audio as that can be done in other packages (e.g. [Librosa](https://librosa.org), [Essentia](https://essentia.upf.edu), [MIR toolbox for Matlab](https://www.jyu.fi/hytk/fi/laitokset/mutku/en/research/materials/mirtoolbox), or [Sonic Visualiser](https://www.sonicvisualiser.org) using well-known onset detection algorithms). Here we take it for granted that the onsets have been extracted already, someone has carried out a quality check, assigned relevant onsets to beat positions and saved them in csv files. For solid overviews on computational onset extraction from audio, see @schluter2014improved or @bock2016madmom, and for a full workflow of how to combine onset detection and annotation of the musical information, see @Clayton2020emr.
 
 ### Onset representation
 
 Here we will take one Cuban Son performance that has been included in the library. This song, _Palo Santo_, has been performed by seven performers using the following combination of instruments: bass, guitar, tres, and trumpet, and five percussion instruments, clave, bongo, bell, cajon, and conga (the instrumentation varies slightly between songs and sections within songs). The full data including other Cuban salsa and son performances is available from an open access repository at _Open Science Framework_ (OSF) and has been annotated and processed as part of the [Interpersonal Entrainment in Music Performance](https://musicscience.net/projects/timing/iemp/) project, available at [https://osf.io/sfxa2/](https://osf.io/sfxa2/). 
 
[…]
