### 2023-10-28 #5395 @tuomaseerola
- [ x] Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper. **Version is 0.5.1**
- [x]  Archive the release on Zenodo/figshare/etc and post the DOI here.  DOI: **10.5281/zenodo.10050346**
- [x]  Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper. **Yes**
- [x]  Make sure that the license listed for the archive is the same as the software license. **Yes**
