+++ paper/paper.bib
@@ -1,7 +1,7 @@
 %% This BibTeX bibliography file was created using BibDesk.
 %% https://bibdesk.sourceforge.io/
 
-%% Created for Tuomas Eerola at 2023-05-28 19:25:24 +0100 
+%% Created for Tuomas Eerola at 2023-10-30 13:31:42 +0000 
 
 
 %% Saved with string encoding Unicode (UTF-8) 
@@ -53,7 +53,8 @@
 @article{goebl2010investigations,
 	author = {Goebl, Werner and Flossmann, Sebastian and Widmer, Gerhard},
 	date-added = {2023-01-17 08:33:33 +0000},
-	date-modified = {2023-01-17 08:33:33 +0000},
+	date-modified = {2023-10-30 13:17:45 +0000},
+	doi = {10.1162/COMJ_a_00002},
 	journal = {Computer Music Journal},
 	number = {3},
 	pages = {35--44},
@@ -85,7 +86,8 @@
 @article{lucas2011inter,
 	author = {Lucas, Glaura and Clayton, Martin and Leante, Laura},
 	date-added = {2023-01-17 08:33:33 +0000},
-	date-modified = {2023-01-17 08:33:33 +0000},
+	date-modified = {2023-10-30 13:20:43 +0000},
+	doi = {10.18061/1811/51203},
 	journal = {Empirical Musicology Review},
 	number = {2},
 	pages = {75--102},
@@ -98,7 +100,8 @@
 	author = {Schl{\"u}ter, Jan and B{\"o}ck, Sebastian},
 	booktitle = {2014 IEEE international conference on acoustics, speech and signal processing (icassp)},
 	date-added = {2023-01-17 08:33:33 +0000},
-	date-modified = {2023-01-17 08:33:33 +0000},
+	date-modified = {2023-10-30 13:21:42 +0000},
+	doi = {10.1109/ICASSP.2014.6854953},
 	organization = {IEEE},
 	pages = {6979--6983},
 	title = {Improved musical onset detection with convolutional neural networks},
@@ -107,7 +110,8 @@
 @article{wallot2018analyzing,
 	author = {Wallot, Sebastian and Leonardi, Giuseppe},
 	date-added = {2023-01-17 08:33:33 +0000},
-	date-modified = {2023-05-28 19:25:24 +0100},
+	date-modified = {2023-10-30 13:22:28 +0000},
+	doi = {10.3389/fpsyg.2018.02232},
 	journal = {Frontiers in Psychology},
 	pages = {2232},
 	publisher = {Frontiers Media SA},
@@ -118,11 +122,11 @@
 @article{wesolowski2016timing,
 	author = {Wesolowski, Brian C},
 	date-added = {2023-01-17 08:33:33 +0000},
-	date-modified = {2023-01-17 08:33:33 +0000},
+	date-modified = {2023-10-30 13:24:16 +0000},
+	doi = {10.1177/0305735614555790},
 	journal = {Psychology of Music},
 	number = {1},
 	pages = {75--94},
-	publisher = {Sage Publications Sage UK: London, England},
 	title = {Timing deviations in jazz performance: The relationships of selected musical variables on horizontal and vertical timing relations: A case study},
 	volume = {44},
 	year = {2016}}
@@ -130,7 +134,8 @@
 @article{lordelo2020adversarial,
 	author = {Lordelo, Carlos and Benetos, Emmanouil and Dixon, Simon and Ahlb{\"a}ck, Sven and Ohlsson, Patrik},
 	date-added = {2023-01-17 08:33:33 +0000},
-	date-modified = {2023-01-17 08:33:33 +0000},
[…]
