### 2023-10-30 #5395 @tuomaseerola
@faroit All added and DOIs double-checked with https://search.crossref.org.
