+++ paper/paper.bib
@@ -1,7 +1,7 @@
 %% This BibTeX bibliography file was created using BibDesk.
 %% https://bibdesk.sourceforge.io/
 
-%% Created for Tuomas Eerola at 2023-10-30 13:31:42 +0000 
+%% Created for Tuomas Eerola at 2023-10-30 14:30:22 +0000 
 
 
 %% Saved with string encoding Unicode (UTF-8) 
@@ -61,7 +61,8 @@
 	publisher = {The MIT Press},
 	title = {Investigations of between-hand synchronization in {M}agaloff's {C}hopin},
 	volume = {34},
-	year = {2010}}
+	year = {2010},
+	bdsk-url-1 = {https://doi.org/10.1162/COMJ_a_00002}}
 
 @article{hawthorne2018enabling,
 	author = {Hawthorne, Curtis and Stasyuk, Andriy and Roberts, Adam and Simon, Ian and Huang, Cheng-Zhi Anna and Dieleman, Sander and Elsen, Erich and Engel, Jesse and Eck, Douglas},
@@ -94,7 +95,8 @@
 	publisher = {Ohio State University, School of Music},
 	title = {Inter-group entrainment in {A}fro-{B}razilian {C}ongado ritual},
 	volume = {6},
-	year = {2011}}
+	year = {2011},
+	bdsk-url-1 = {https://doi.org/10.18061/1811/51203}}
 
 @inproceedings{schluter2014improved,
 	author = {Schl{\"u}ter, Jan and B{\"o}ck, Sebastian},
@@ -105,7 +107,8 @@
 	organization = {IEEE},
 	pages = {6979--6983},
 	title = {Improved musical onset detection with convolutional neural networks},
-	year = {2014}}
+	year = {2014},
+	bdsk-url-1 = {https://doi.org/10.1109/ICASSP.2014.6854953}}
 
 @article{wallot2018analyzing,
 	author = {Wallot, Sebastian and Leonardi, Giuseppe},
@@ -117,7 +120,8 @@
 	publisher = {Frontiers Media SA},
 	title = {Analyzing multivariate dynamics using cross-recurrence quantification analysis ({CRQA}), diagonal-cross-recurrence profiles (dcrp), and multidimensional recurrence quantification analysis ({MDRQA})-- {A} tutorial in {R}},
 	volume = {9},
-	year = {2018}}
+	year = {2018},
+	bdsk-url-1 = {https://doi.org/10.3389/fpsyg.2018.02232}}
 
 @article{wesolowski2016timing,
 	author = {Wesolowski, Brian C},
@@ -129,7 +133,8 @@
 	pages = {75--94},
 	title = {Timing deviations in jazz performance: The relationships of selected musical variables on horizontal and vertical timing relations: A case study},
 	volume = {44},
-	year = {2016}}
+	year = {2016},
+	bdsk-url-1 = {https://doi.org/10.1177/0305735614555790}}
 
 @article{lordelo2020adversarial,
 	author = {Lordelo, Carlos and Benetos, Emmanouil and Dixon, Simon and Ahlb{\"a}ck, Sven and Ohlsson, Patrik},
@@ -141,7 +146,8 @@
 	publisher = {IEEE},
 	title = {Adversarial unsupervised domain adaptation for harmonic-percussive source separation},
 	volume = {28},
-	year = {2020}}
+	year = {2020},
+	bdsk-url-1 = {https://doi.org/10.1109/LSP.2020.3045915}}
 
 @inproceedings{sordo2014creating,
 	author = {Sordo, Mohamed and Chaachoo, Amin and Serra, Xavier},
@@ -151,7 +157,8 @@
 	doi = {10.1145/2660168.2660182},
 	pages = {1--3},
[…]
