### 2023-10-30 #5395 @tuomaseerola
@faroit Good, thanks, it is actually 10.3758/BF03206433 or https://doi.org/10.3758/bf03206433, I missed the 2005 paper in the first pass as Bruno Repp has nearly the same title in his 2013 paper. Now added.
