### 2023-11-06 #5395 @tuomaseerola
@faroit any idea why the bot initiated process hangs in the middle (the paper is compiled successfully during the process)?
