### 2023-11-17 #5395 @arfon
@tuomaseerola  – `Clayton2020emr` is defined twice in your BibTeX file. Could you verify they are duplicates and remove one if that is the case?
