+++ paper/paper.bib
@@ -1,7 +1,7 @@
 %% This BibTeX bibliography file was created using BibDesk.
 %% https://bibdesk.sourceforge.io/
 
-%% Created for Tuomas Eerola at 2023-10-30 14:30:22 +0000 
+%% Created for Tuomas Eerola at 2023-11-21 21:09:08 +0000 
 
 
 %% Saved with string encoding Unicode (UTF-8) 
@@ -244,7 +244,8 @@
 	publisher = {Springer},
 	title = {Sensorimotor synchronization: A review of the tapping literature},
 	volume = {12},
-	year = {2005}}
+	year = {2005},
+	bdsk-url-1 = {https://doi.org/10.3758/BF03206433}}
 
 @article{clayton_et_al_2018b,
 	author = {Clayton, M. and Jakubowski, K. and Eerola, T.},
@@ -259,21 +260,6 @@
 	volume = {23},
 	year = {2019},
 	bdsk-url-1 = {https://doi.org/10.1177/1029864919844809}}
-
-@article{Clayton2020emr,
-	author = {Clayton, M and Tarsitani, S and Jankowsky, R and Jure L and Leante, L and Polak, R and Poole, A and Rocamora, M and Alborno, P and Camurri, A and Eerola, T and Jacoby, N and Jakubowski, K},
-	date-added = {2023-01-17 08:33:33 +0000},
-	date-modified = {2023-01-17 08:33:33 +0000},
-	doi = {10.18061/emr.v16i1.7555},
-	journal = {Empirical Musicology Review},
-	number = {1},
-	pages = {65-84},
-	title = {The Interpersonal Entrainment in Music Performance Data Collection},
-	url = {10.18061/emr.v16i1.7555},
-	volume = {16},
-	year = {2021},
-	bdsk-url-1 = {10.18061/emr.v16i1.7555},
-	bdsk-url-2 = {https://doi.org/10.18061/emr.v16i1.7555}}
 
 @article{rasch1979synchronization,
 	author = {Rasch, Rudolf A},
