### 2023-11-21 #5395 @tuomaseerola
@arfon Thanks for the pointer, it was indeed a duplicate and I have removed it from the BibTeX file (with new commit/push).
