### 2023-12-07 #5395 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
