### 2024-01-06 #5395 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
