### 2024-01-06 #5395 @oliviaguest
Huge thanks to the editor: @faroit and the reviewers: @pmcharrison, @mrocamora! ✨  JOSS appreciates your work and effort. ✨ Also, big congratulations to the authors: @tuomaseerola! 🥳 🍾
